完善功能：
增加用户获取广告信息接口以及其测试

增加管理员登出接口

+++ src/main/java/com/adv/pojo/Administrator.java
@@ -1,14 +1,12 @@
 package com.adv.pojo;
 
-import java.util.List;
-
 /**
  * @author lurongzhi
  */
 public class Administrator {
     private String name;
     private String password;
-    private Integer id;
+    private Long id;
     private String nextUrl;
 
 
@@ -21,11 +19,11 @@ public class Administrator {
         this.nextUrl = nextUrl;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
